+++ src/topi/transform.cc
@@ -24,6 +24,7 @@
 #include <tvm/runtime/packed_func.h>
 #include <tvm/runtime/registry.h>
 #include <tvm/topi/transform.h>
+#include <tvm/topi/einsum.h>
 #include <tvm/topi/utils.h>
 
 namespace tvm {
